moved capsules Table UI to content View
